requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.26.0
redis==5.0.1
python-dateutil==2.8.2
PyYAML==6.0.1
pillow==10.2.0
email-validator==2.1.0
phonenumbers==8.13.27
